+++ packages/phrases/src/languages/en/translation/admin-console/api-resource-details.ts
