+++ packages/phrases/src/languages/en/translation/admin-console/api-resources.ts
