+++ packages/phrases/src/languages/en/translation/admin-console/application-details.ts
