+++ packages/phrases/src/languages/en/translation/admin-console/applications.ts
