+++ packages/phrases/src/languages/en/translation/admin-console/connector-details.ts
