+++ packages/phrases/src/languages/en/translation/admin-console/connectors.ts
