+++ packages/phrases/src/languages/en/translation/admin-console/dashboard.ts
