+++ packages/phrases/src/languages/en/translation/admin-console/general.ts
