+++ packages/phrases/src/languages/en/translation/admin-console/get-started.ts
