+++ packages/phrases/src/languages/en/translation/admin-console/log-details.ts
