+++ packages/phrases/src/languages/en/translation/admin-console/logs.ts
