+++ packages/phrases/src/languages/en/translation/admin-console/settings.ts
