+++ packages/phrases/src/languages/en/translation/admin-console/tab-sections.ts
