+++ packages/phrases/src/languages/en/translation/admin-console/tabs.ts
