+++ packages/phrases/src/languages/en/translation/admin-console/user-details.ts
