+++ packages/phrases/src/languages/en/translation/admin-console/welcome.ts
