+++ packages/phrases/src/languages/zh-cn/index.ts
