+++ packages/phrases/src/languages/zh-cn/translation/admin-console/api-resource-details.ts
