+++ packages/phrases/src/languages/zh-cn/translation/admin-console/api-resources.ts
