+++ packages/phrases/src/languages/zh-cn/translation/admin-console/application-details.ts
