+++ packages/phrases/src/languages/zh-cn/translation/admin-console/connector-details.ts
