+++ packages/phrases/src/languages/zh-cn/translation/admin-console/general.ts
