+++ packages/phrases/src/languages/zh-cn/translation/admin-console/get-started.ts
