+++ packages/phrases/src/languages/zh-cn/translation/admin-console/index.ts
