+++ packages/phrases/src/languages/zh-cn/translation/admin-console/log-details.ts
