+++ packages/phrases/src/languages/zh-cn/translation/admin-console/logs.ts
