+++ packages/phrases/src/languages/zh-cn/translation/admin-console/tab-sections.ts
