+++ packages/phrases/src/languages/zh-cn/translation/admin-console/tabs.ts
