+++ packages/phrases/src/languages/zh-cn/translation/admin-console/user-details.ts
