+++ packages/phrases/src/languages/zh-cn/translation/demo-app.ts
