+++ packages/phrases/src/languages/zh-cn/translation/index.ts
