+++ packages/phrases/src/locales/en.ts
@@ -1,5 +1,5 @@
 /**
  * Note: we need to re-export the locale under this `locales` folder,
- * so that the `useTranslation` function can infer the i18n key type.
+ * otherwise the `useTranslation` function will cause a `Type instantiation` error (ts2589).
  */
-export { default } from '../language/en';
+export { default } from '../languages/en';
